Prevent removing an organisation's default brand
Removing the default brand from an organisation's available brands
would leave the default pointing at a brand the organisation can no
longer use.

Refuse to destroy the organisation brand when it is the current
default and redirect back to the organisation page. The remove button
will be disabled in the interface for the default brand, so this guard
only protects against crafted requests.

+++ app/controllers/organisation_brands_controller.rb
@@ -23,6 +23,11 @@ class OrganisationBrandsController < WebController
     authorize organisation, :can_manage_organisation_brands?
 
     organisation_brand = organisation.organisation_brands.find_by!(brand_id: params[:id])
+
+    if organisation.default_brand_id == organisation_brand.brand_id
+      return redirect_to organisation_path(organisation), status: :see_other
+    end
+
     organisation_brand.destroy!
 
     redirect_to organisation_path(organisation), success: t(".success", brand_name: organisation_brand.brand.name)
